[EI-127] [web] refactor: Refactor store reset functionality
