+++ web/__tests__/components/numerical-guidance/indicator-list.test.tsx
@@ -3,7 +3,8 @@ import userEvent from '@testing-library/user-event';
 import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
 import { resetMockDB } from '@/app/mocks/db.mock';
 import IndicatorList from '@/app/ui/components/numerical-guidance/indicator-list';
-import { resetAllStore, useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance.store';
+import { useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance.store';
+import { resetAllStore } from '@/app/store/stores/reset-store';
 
 describe('IndicatorList', () => {
   beforeEach(() => {
