+++ web/__tests__/hooks/use-indicator-metadata-list.hook.test.tsx
@@ -3,7 +3,8 @@ import { renderHook, waitFor } from '@testing-library/react';
 import { server } from '@/app/mocks/server.mock';
 import { HttpResponse, http } from 'msw';
 import { API_PATH } from '@/app/store/querys/api-path';
-import { resetAllStore, useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance.store';
+import { useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance.store';
+import { resetAllStore } from '@/app/store/stores/reset-store';
 import { act } from 'react-dom/test-utils';
 import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
 import { resetMockDB } from '@/app/mocks/db.mock';
