+++ web/__tests__/hooks/use-indicators-value-view-model.hook.test.tsx
@@ -1,7 +1,8 @@
 import { act, renderHook, waitFor } from '@testing-library/react';
 import { SWRProviderWithoutCache } from '@/app/store/querys/swr-provider';
 import { resetMockDB, mockDB } from '@/app/mocks/db.mock';
-import { resetAllStore, useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance.store';
+import { useNumericalGuidanceStore } from '@/app/store/stores/numerical-guidance.store';
+import { resetAllStore } from '@/app/store/stores/reset-store';
 import { useIndicatorsValueViewModel } from '@/app/business/hooks/use-indicators-value-view-model.hook';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/use-selected-indicator-board-metadata.hook';
 import { IndicatorsValue } from '@/app/business/services/indicators-value-view-model.service';
