+++ web/app/store/stores/reset-store.ts
@@ -1,0 +1,7 @@
+export const storeResetFns = new Set<() => void>();
+
+export const resetAllStore = () => {
+  storeResetFns.forEach((resetFn) => {
+    resetFn();
+  });
+};
